fix fluidstack and paperspace
